requirements.txt
# CLI chat app
github-copilot-sdk
asyncpg

# Dashboard
fastapi
uvicorn[standard]
jinja2
